pyproject.toml
[tool.poetry]
name = "multi-tools-pdf-extractor"
version = "0.1.0"
description = ""
authors = ["hallan-n <hallanneves01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
httpx = "^0.27.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.28"}
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
cryptography = "^42.0.5"
uvicorn = "^0.28.0"
python-multipart = "^0.0.9"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.1.1"
pre-commit = "^3.6.2"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
